requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
httpx>=0.24.0
loguru>=0.7.0
sqlalchemy>=2.0.0

# Frontend & Client
streamlit>=1.24.0
requests>=2.31.0
pandas>=2.0.0
graphviz>=0.20.1

# RAG & ML
chromadb>=0.4.0
sentence-transformers>=2.2.2

# Constraints for compatibility
numpy<2.0.0
protobuf<5.0.0
packaging<24.0
tenacity<9.0.0
# Force opencv to be compatible with numpy<2
opencv-python-headless<4.10.0.0
